+++ src/wujihandpy/__init__.py
@@ -1,6 +1,10 @@
 from __future__ import annotations
 
 from ._core import *
-from ._version import __version__
+
+try:
+    from ._version import __version__
+except ImportError:
+    __version__ = "0+unknown"
 
 __all__ = ["__version__", 'Hand', 'Finger', 'Joint', 'IController', 'filter']